requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
transformers==4.38.0
torch==2.2.0
scikit-learn==1.4.0
joblib==1.3.2
openai==1.12.0
faiss-cpu==1.7.4
sentence-transformers==2.3.0
spacy==3.7.2
python-dotenv==1.0.1
pytest==8.0.0
